+++ src/mutations/remove_node.rs
@@ -15,7 +15,7 @@ impl Mutations {
             // make iterator wrap
             .cycle()
             // randomly offset into the iterator to choose any node
-            .skip((rng.gen::<f64>() * (genome.hidden.len()) as f64).floor() as usize)
+            .skip((rng.gen::<f32>() * (genome.hidden.len()) as f32).floor() as usize)
             // just loop every value once
             .take(genome.hidden.len())
             .find(|removal_candidate| {
